requirements.txt
Pillow==10.2.0
img2pdf==0.5.1 